Extract dashboard static assets
